WIP move ellipsis menu to its own file and put it in the top of the post card
