+++ client/blocks/reader-post-card/byline.jsx
@@ -4,6 +4,7 @@ import PropTypes from 'prop-types';
 import { Component } from 'react';
 import ReaderAuthorLink from 'calypso/blocks/reader-author-link';
 import ReaderAvatar from 'calypso/blocks/reader-avatar';
+import ReaderPostEllipsisMenu from 'calypso/blocks/reader-post-options-menu/reader-post-ellipsis-menu';
 import ReaderSiteStreamLink from 'calypso/blocks/reader-site-stream-link';
 import TimeSince from 'calypso/components/time-since';
 import { areEqualIgnoringWhitespaceAndCase } from 'calypso/lib/string';
@@ -53,6 +54,8 @@ class PostByline extends Component {
 		isDiscoverPost: PropTypes.bool,
 		showSiteName: PropTypes.bool,
 		showAvatar: PropTypes.bool,
+		teams: PropTypes.array,
+		showFollow: PropTypes.bool,
 	};
 
 	static defaultProps = {
@@ -69,7 +72,8 @@ class PostByline extends Component {
 	};
 
 	render() {
-		const { post, site, feed, isDiscoverPost, showSiteName, showAvatar } = this.props;
+		const { post, site, feed, isDiscoverPost, showSiteName, showAvatar, teams, showFollow } =
+			this.props;
 		const feedId = get( post, 'feed_ID' );
 		const siteId = get( site, 'ID' );
 		const siteSlug = get( site, 'slug' );
@@ -163,6 +167,12 @@ class PostByline extends Component {
 						) }
 					</div>
 				</div>
+				<ReaderPostEllipsisMenu
+					site={ site }
+					teams={ teams }
+					post={ post }
+					showFollow={ showFollow }
+				/>
 			</div>
 		);
 		/* eslint-enable wpcalypso/jsx-gridicon-size */
